simplify source unit data structures
